RESTEASY-1246 Part II
